# src/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Brazilian Financial Data API
  description: |
    Unified API for accessing Brazilian financial market data from public sources.

    ## Services
    - **CVM** (port 8000): Credit market data - FIDC, FIP, FIAGRO, SECURIT
    - **BACEN** (port 8002): Central Bank data - SGS, PTAX, Focus expectations
    - **B3 CALC** (port 8001): Fixed income pricing - debentures, CRA, CRI

    ## Authentication
    All endpoints are publicly accessible (no authentication required).
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/PedroDnT/iliquid_nightly
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server (CVM API)
  - url: http://localhost:8001
    description: Local development server (B3 CALC API)
  - url: http://localhost:8002
    description: Local development server (BACEN API)
...
paths:
...
  /api/v1/bacen/sgs/{series_code}:
    get:
      tags:
        - BACEN
      summary: Get SGS time series
      description: Retrieve a specific time series from BACEN SGS (Sistema Gerenciador de Séries)
      operationId: getSgsSeries
      parameters:
        - name: series_code
          in: path
          required: true
          schema:
            type: integer
          example: 432
        - name: data_inicio
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: data_fim
          in: query
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SGSSeriesResponse"
  /api/v1/bacen/sgs/multi:
    post:
      tags:
        - BACEN
      summary: Get multiple SGS series
      description: Retrieve multiple time series in a single request
      operationId: getMultiSgsSeries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiSgsRequest"
      responses:
        "200":
          description: Multiple time series data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiSgsResponse"
  /api/v1/bacen/ptax/dolar:
    get:
      tags:
        - BACEN
      summary: Get USD/BRL exchange rate
      description: Retrieve USD to BRL exchange rate from PTAX
      operationId: getPtaxDolar
      parameters:
        - name: data
          in: query
          description: Date (YYYY-MM-DD). If not provided, returns latest.
          schema:
            type: string
            format: date
          example: "2024-01-15"
      responses:
        "200":
          description: Exchange rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXRateResponse"
  /api/v1/bacen/ptax/dolar/periodo:
    get:
      tags:
        - BACEN
      summary: Get USD/BRL exchange rate for period
      description: Retrieve USD to BRL exchange rate for a date range
      operationId: getPtaxDolarPeriod
      parameters:
        - name: data_inicio
          in: query
          required: true
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-01-01"
        - name: data_fim
          in: query
          required: true
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
          example: "2024-12-31"
      responses:
        "200":
          description: Exchange rate data for period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXPeriodResponse"
...
  /api/v1/bacen/ptax/moeda/{moeda}:
    get:
      tags:
        - BACEN
      summary: Get exchange rate for specific currency
      description: Retrieve exchange rate for any available currency
      operationId: getPtaxMoeda
      parameters:
        - name: moeda
          in: path
          required: true
          description: Currency code (e.g., EUR, GBP, JPY)
          schema:
            type: string
          example: EUR
        - name: data
          in: query
          description: Date (YYYY-MM-DD). If not provided, returns latest.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Exchange rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTAXRateResponse"
...
  /api/v1/bacen/expectativas/{endpoint_name}:
    get:
      tags:
        - BACEN
      summary: Get market expectations
      description: Query Focus/RD Mercado market expectations data
      operationId: getExpectativas
      parameters:
        - name: endpoint_name
          in: path
          required: true
          schema:
            type: string
            enum:
              - ExpectativasMercadoAnuais
              - ExpectativasMercadoMensais
              - ExpectativasMercadoTrimestrais
              - ExpectativasMercadoSelic
              - ExpectativasMercadoTop5Anuais
          example: ExpectativasMercadoMensais
        - name: indicador
          in: query
          schema:
            type: string
        - name: data
          in: query
          schema:
            type: string
            format: date
        - name: topx
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Market expectations data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpectativasResponse"
  /api/v1/bacen/taxas_juros/{endpoint_name}:
    get:
      tags:
        - BACEN
      summary: Get interest rate data
      description: Query BACEN interest rate OData endpoints
      operationId: getTaxasJuros
      parameters:
        - name: endpoint_name
          in: path
          required: true
          schema:
            type: string
          example: TaxaJurosDiaria
      responses:
        "200":
          description: Interest rate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxaJurosResponse"
...
components:
  schemas:
...
    SGSSeriesResponse:
      type: object
      properties:
        series_code:
          type: integer
        series_name:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              data:
                type: string
              valor:
                type: number
    MultiSgsRequest:
      type: object
      properties:
        series:
          type: array
          items:
            type: integer
        data_inicio:
          type: string
          format: date
        data_fim:
          type: string
          format: date
    MultiSgsResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/SGSSeriesResponse"
    PTAXRateResponse:
      type: object
      properties:
        moeda:
          type: string
        data:
          type: string
        cotacao_venda:
          type: number
        cotacao_compra:
          type: number
    PTAXPeriodResponse:
      type: object
      properties:
        moeda:
          type: string
        data_inicio:
          type: string
        data_fim:
          type: string
        dados:
          type: array
          items:
            type: object
...
    ExpectativasResponse:
      type: object
      properties:
        endpoint:
          type: string
        data:
          type: array
          items:
            type: object
...
    TaxaJurosResponse:
      type: object
      properties:
        endpoint:
          type: string
        data:
          type: array
          items:
            type: object
...